+++ src/GameLogic/Rules/DeathmatchRules.cs
@@ -20,6 +20,7 @@ public class DeathmatchRules : IBattleRules
     public bool IsFriendlyFireEnabled => true;
     public bool UseShrinkingZone => false;
     public bool UsesPowerups => true;
+    public int MinPlayersToStart => Constants.MinPlayersToStart;
 
     public void Initialize(GameRoomState state)
     {